+++ client/src/app/shared/types/card-creation.types.ts
@@ -1,53 +1,6 @@
-import { CardData } from '../../parser/types';
+import { CardType } from '../../parser/types';
 
-export type CardType = 'vocabulary';
-
-export interface ExtractedItem {
-  id: string;
-  exists: boolean;
-}
-
-export interface ExtractionRequest {
-  sourceId: string;
-  pageNumber: number;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-export interface ExtractionRegion {
-  rectangle: { x: number; y: number; width: number; height: number };
-  items: ExtractedItem[];
-}
-
-export interface CardCreationRequest {
-  item: ExtractedItem;
-  sourceId: string;
-  pageNumber: number;
-  cardType: CardType;
-}
-
-export interface ImageGenerationInfo {
-  cardId: string;
-  exampleIndex: number;
-  englishTranslation: string;
-}
-
-export interface CardCreationResult {
-  cardData: CardData;
-  imageGenerationInfos: ImageGenerationInfo[];
-}
-
-export interface CardCreationStrategy {
-  cardType: CardType;
-  extractItems(request: ExtractionRequest): Promise<ExtractedItem[]>;
-  getItemLabel(item: ExtractedItem): string;
-  filterItemsBySearchTerm(items: ExtractedItem[], searchTerm: string): ExtractedItem[];
-  createCardData(request: CardCreationRequest, progressCallback: (progress: number, step: string) => void): Promise<CardCreationResult>;
-}
-
-export interface CardCreationProgress {
+export type CardCreationProgress = {
   itemLabel: string;
   cardType: CardType;
   status: 'pending' | 'word-type' | 'translating' | 'generating-images' | 'creating-card' | 'completed' | 'error';
